There are bullets!

+++ main/modes/games/danceNetwork/dn_utility.c
@@ -57,7 +57,7 @@ void dn_setFloorPalette(wsgPalette_t* palette, paletteColor_t color)
  * @param b One of two inputs
  * @param amount Lerp amount from 0 to 30000. 0 returns a, 30000 returns b.
  */
-int16_t dn_lerp(int16_t a, int16_t b, uint16_t amount)
+int dn_lerp(int a, int b, uint16_t amount)
 {
     return a + ((b - a) * amount) / 30000;
 }
