+++ main/modes/games/danceNetwork/dn_utility.h
@@ -12,6 +12,6 @@ dn_assetIdx_t dn_getAssetIdx(dn_characterSet_t characterSet, dn_unitRank rank, d
 
 void dn_setFloorPalette(wsgPalette_t* palette, paletteColor_t color);
 
-int16_t dn_lerp(int16_t a, int16_t b, uint16_t amount);
+int dn_lerp(int a, int b, uint16_t amount);
 
 int16_t dn_logRemap(int16_t x);
